+++ hmm/lib/c/hmm.h
@@ -7,14 +7,14 @@ double forward(
 		const double *A,
 		const double *B,
 		const double *pi,
-		const int *O,
+		const short *O,
 		int N, int M, int T);
 
 void backward(
 		double *beta,
 		const double *A,
 		const double *B,
-		const int *O,
+		const short *O,
 		const double *scale,
 		int N, int M, int T);
 
@@ -22,7 +22,7 @@ void compute_nomA(
 		double *nomA,
 		const double *A,
 		const double *B,
-		const int *O,
+		const short *O,
 		const double *alpha,
 		const double *beta,
 		int N, int M, int T);
@@ -35,7 +35,7 @@ void compute_denomA(
 void compute_nomB(
 		double *nomB,
 		const double *gamma,
-		const int *O,
+		const short *O,
 		int N, int M, int T);
 
 void computeGamma(
@@ -48,7 +48,7 @@ void computeXi(
 		double *xi,
 		const double *A,
 		const double *B,
-		const int *O,
+		const short *O,
 		const double *alpha,
 		const double *beta,
 		int N, int M, int T);
@@ -57,7 +57,7 @@ void update(
 		double *A,
 		double *B,
 		double *pi,
-		const int *O,
+		const short *O,
 		const double *gamma,
 		const double *xi,
 		int N, int M, int T);
@@ -78,14 +78,14 @@ float forward32(
 		const float *A,
 		const float *B,
 		const float *pi,
-		const int *O,
+		const short *O,
 		int N, int M, int T);
 
 void backward32(
 		float *beta,
 		const float *A,
 		const float *B,
-		const int *O,
+		const short *O,
 		const float *scale,
 		int N, int M, int T);
 
@@ -93,7 +93,7 @@ void compute_nomA32(
 		float *nomA,
 		const float *A,
 		const float *B,
-		const int *O,
+		const short *O,
 		const float *alpha,
 		const float *beta,
 		int N, int M, int T);
@@ -106,7 +106,7 @@ void compute_denomA32(
 void compute_nomB32(
 		float *nomB,
 		const float *gamma,
-		const int *O,
+		const short *O,
 		int N, int M, int T);
 
 void computeGamma32(
@@ -119,7 +119,7 @@ void computeXi32(
 		float *xi,
 		const float *A,
 		const float *B,
-		const int *O,
+		const short *O,
 		const float *alpha,
 		const float *beta,
 		int N, int M, int T);
@@ -128,7 +128,7 @@ void update32(
 		float *A,
 		float *B,
 		float *pi,
-		const int *O,
+		const short *O,
 		const float *gamma,
 		const float *xi,
 		int N, int M, int T);
